+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+### Added
+- Added TAG_TOOL_PTK_PASSIVE_DEFAULT, TAG_TOOL_PTK_ACTIVE_DEFAULT tags.
+
 ## 0.9.0 - 2026-09-01
 ### Changed
 - Update the bundled OWASP Penetration Testing Kit browser extensions to the signed 9.9.9 store artifacts.